[CEEZ-63] refactor: 다운로드, 좋아요 관련 로그 albumLogger로 이동
